requirements.txt
google
bs4
requests
pytest
html5lib
flask
Flask-SQLAlchemy
pandas
awscli
click
scrapy
pydantic
rich
streamlit
